further move stuff around and split it up
