+++ corehq/ex-submodules/casexml/apps/phone/tests/performance_tests.py
@@ -1,10 +1,12 @@
 from casexml.apps.case.mock import CaseBlock
-from casexml.apps.phone.caselogic import filter_cases_modified_elsewhere_since_sync, get_related_cases
+from casexml.apps.phone.caselogic import get_related_cases
+from casexml.apps.phone.data_providers import get_case_payload_batched
+from casexml.apps.phone.data_providers.case.batched import filter_cases_modified_elsewhere_since_sync
 from casexml.apps.phone.tests.test_sync_mode import SyncBaseTest, PARENT_TYPE
 from casexml.apps.phone.tests.utils import synclog_from_restore_payload, generate_restore_payload
 from casexml.apps.case.tests.util import assert_user_has_cases
 from casexml.apps.phone.models import User
-from casexml.apps.phone.restore import RestoreConfig, get_case_payload_batched
+from casexml.apps.phone.restore import RestoreConfig
 from dimagi.utils.decorators.profile import line_profile
 from casexml.apps.case.xml import V2
 from datetime import datetime
